requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16,<3.17
django-cors-headers>=4.6,<5.1
argon2-cffi>=23.1,<26.0
gunicorn>=23.0,<24.0
psycopg[binary]>=3.2,<3.3
meilisearch>=0.34,<0.35
openai>=1.109,<2.0
pluggy>=1.5,<2.0
stripe>=7.0,<12.0
boto3>=1.35,<2.0
django-storages>=1.14,<2.0
sentry-sdk[django]>=2.0,<3.0
celery>=5.3.4,<6.0
redis>=5.0.0,<6.0
pre-commit>=3.6,<4.0
